LPS-168771 - Move ClayVerticalNavItem to inner class of EditRolePermissionsNavigationDisplayContext.
