+++ modules/apps/roles/roles-admin-web/src/main/java/com/liferay/roles/admin/web/internal/display/context/EditRolePermissionsNavigationDisplayContext.java
@@ -14,6 +14,9 @@
 
 package com.liferay.roles.admin.web.internal.display.context;
 
+import com.fasterxml.jackson.annotation.JsonAnyGetter;
+import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
 import com.liferay.application.list.PanelApp;
@@ -36,6 +39,7 @@ import com.liferay.portal.kernel.portlet.OmniadminControlPanelEntry;
 import com.liferay.portal.kernel.portlet.PortletProvider;
 import com.liferay.portal.kernel.portlet.PortletProviderUtil;
 import com.liferay.portal.kernel.portlet.url.builder.ResourceURLBuilder;
+import com.liferay.portal.kernel.security.SecureRandomUtil;
 import com.liferay.portal.kernel.service.PortletLocalServiceUtil;
 import com.liferay.portal.kernel.theme.ThemeDisplay;
 import com.liferay.portal.kernel.util.ArrayUtil;
@@ -52,7 +56,10 @@ import com.liferay.product.navigation.personal.menu.BasePersonalMenuEntry;
 import com.liferay.roles.admin.constants.RolesAdminWebKeys;
 
 import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collection;
 import java.util.Collections;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Locale;
 import java.util.Map;
@@ -517,5 +524,78 @@ public class EditRolePermissionsNavigationDisplayContext {
 	private final Role _role;
 	private final ServletContext _servletContext;
 	private final ThemeDisplay _themeDisplay;
+
+	@JsonInclude(JsonInclude.Include.NON_DEFAULT)
+	private static class ClayVerticalNavItem {
+
+		public static ClayVerticalNavItem create(
+			String label,
+			Consumer<ClayVerticalNavItem> clayVerticalNavItemConsumer) {
+
+			ClayVerticalNavItem clayVerticalNavItem = new ClayVerticalNavItem(
+				label);
+
+			clayVerticalNavItemConsumer.accept(clayVerticalNavItem);
+
+			if (clayVerticalNavItem.id == null) {
+				clayVerticalNavItem.setId(
+					_CLAY_VERTICAL_NAV_ITEM + SecureRandomUtil.nextLong());
+			}
+
+			return clayVerticalNavItem;
+		}
+
+		public ClayVerticalNavItem(String label) {
+			this.label = label;
+		}
+
+		public void addItems(ClayVerticalNavItem... clayVerticalNavItems) {
+			addItems(Arrays.asList(clayVerticalNavItems));
+		}
+
+		public void addItems(
+			Collection<ClayVerticalNavItem> clayVerticalNavItems) {
+
+			items.addAll(clayVerticalNavItems);
+		}
+
+		public void put(String key, Object value) {
+			properties.put(key, value);
+		}
+
+		public void setActive(boolean active) {
+			this.active = active;
+		}
+
+		public void setId(String id) {
+			this.id = id;
+		}
+
+		public void setInitialExpanded(boolean initialExpanded) {
+			this.initialExpanded = initialExpanded;
+		}
+
+		@JsonProperty
+		protected boolean active;
+
+		@JsonProperty
+		protected String id;
+
+		@JsonProperty
+		protected boolean initialExpanded;
+
+		@JsonProperty
+		protected List<ClayVerticalNavItem> items = new ArrayList<>();
+
+		@JsonProperty
+		protected final String label;
+
+		@JsonAnyGetter
+		protected final Map<String, Object> properties = new HashMap<>();
+
+		private static final String _CLAY_VERTICAL_NAV_ITEM =
+			"CLAY_VERTICAL_NAV_ITEM";
+
+	}
 
 }
